FAME now accepts the numbers of isomers, reactants, and products from input.
Before it would read these numbers and use them to process the input file,
then oddly attempt to recompute them. There were a few assumptions in this
recomputation that are about to be no longer valid. It is better to just
accept what the input file tells you they are anyway.