Now allowing bimolecular channels to be product channels in PDep calculations.
In order to be a bimolecular reactant channel, both species in the
configuration must now be in the model core (i.e. so that they might have
nonzero concentration). This generally means fewer reactant channels, leading
to slightly faster FAME calculations. However, this will also require more
calls to FAME, e.g. when a bimolecular product channel becomes a bimolecular
reactant channel as species are added to the core.