Significant changes to how PDepNetworks are explored.
First, each network now has a designated source configuration, and is
responsible for producing rate coefficients for that configuration reacting
to produce anything else in that network. This generally results in many more
networks and hence many more calls to FAME; however, each individual call
should run a bit faster than before, countering the speed penalty somewhat.

The benefit of this approach is that we avoid exploring networks except for
those that really need it. Before, the procedure required merging of networks
that contained any common included isomers. This had a tendency to create
very large networks in the course of huge RMG runs. These very large networks
often contained a lot of unnecessary high-energy configurations, which was a
waste of time. Furthermore, the output from FAME for these networks was so
verbose that it would cause Java's BufferedReader to "clog" and take
unusually long to process.

Merging of networks has been completely removed at the moment. However, I
expect that we will need to do some merging of networks that have the same
source and common unimolecular isomers. This has not been implemented yet.

The static method PDepNetwork.addReactionToNetworks() is now to be used only
when a species is being added to the model core. The non-static method
PDepNetwork.makeIsomerIncluded() no longer calls this method, since all
generated reactions are automatically added to that network, and we're not
interested in any merging at this time.