Mark all networks containing a new core species as needing an update.
Any PDepNetwork containing the new core species in any configuration -
unimolecular or bimolecular - is potentially in need of an update. There are
some special circumstances wherein the network might not need an update, but
we will not code those yet. This is necessary because the only source terms
allowed in each FAME call are those in the current core. If the core changes,
new terms in the network might become source terms that weren't before.