Restored merging of partial networks under certain specific circumstances.
We need to merge networks that have the same entrant configuration and that
share at least one other included isomer. If we don't, there will be two
partial networks that produce that reaction as a net reaction in the same
direction, so we will be double counting. Also, this merging is consistent
with the fact that they are in fact one network.

Note that this is still less merging of networks overall than the previous
algorithm, which would merge all networks containing an isomer as soon as
that isomer became included in *any* network. The hope is that this new
algorithm will avoid the gigantic networks that are encountered when zealous
merging is used.