Now treating all PDepReactions as irreversible.
This is one of the outcomes of the switch to minimize merging of networks.
Each partial network is now responsible for producing reactions from its
source to all other configurations in that network. There will usually be
another network that provides the reverse reaction.

Note that this does not guarantee thermodynamic consistency, though we'll
probably still be pretty close.